requirements.txt
Cython
dataclasses
GitPython
greenlet
h5py
hdfs
hyperopt
imageio
imbalanced-learn
joblib
jsoncomment
llvmlite
lxml
matplotlib
numba
numpy
opencv-contrib-python
openpyxl
pandas
Pillow
plotly
protobuf
pydicom
PyTexturePacker
scikit-image
scikit-learn
scikit-optimize
scipy
seaborn
xgboost
